backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery[redis]==5.3.6
redis==5.0.1
spacy==3.7.2
openpyxl==3.1.2
pandas==2.1.4
httpx==0.26.0
defusedxml==0.7.1
pytest==7.4.4
pytest-asyncio==0.23.3
